feat: add production environment setup with environment-aware seeding
- Create production seed script for clean production setup (admin + shift types + templates only)
- Add environment-aware seed system using VERCEL_ENV as source of truth
- Reorganize seed files: seed.js (auto-detect), seed-demo.js (full demo), seed-production.js (minimal)
- Hide quick login demo credentials in production environments
- Add comprehensive shift types and templates for all locations (Wellington, Glen Innes, Onehunga)
- Configure Prisma to automatically use appropriate seed based on environment
